+++ lithic-java-core/src/main/kotlin/com/lithic/api/models/AccountHolderSimulateEnrollmentDocumentReviewResponse.kt
@@ -18,22 +18,24 @@ import com.lithic.api.errors.LithicInvalidDataException
 import java.util.Objects
 import java.util.Optional
 
-/** Describes the document and the required document image uploads required to re-run KYC. */
-@JsonDeserialize(builder = AccountHolderDocument.Builder::class)
+/** A document to be submitted for account holder verification. */
+@JsonDeserialize(builder = AccountHolderSimulateEnrollmentDocumentReviewResponse.Builder::class)
 @NoAutoDetect
-class AccountHolderDocument
+class AccountHolderSimulateEnrollmentDocumentReviewResponse
 private constructor(
+    private val token: JsonField<String>,
     private val accountHolderToken: JsonField<String>,
     private val documentType: JsonField<DocumentType>,
-    private val entityToken: JsonField<String>,
     private val requiredDocumentUploads: JsonField<List<RequiredDocumentUpload>>,
-    private val token: JsonField<String>,
     private val additionalProperties: Map<String, JsonValue>,
 ) {
 
     private var validated: Boolean = false
 
     private var hashCode: Int = 0
+
+    /** Globally unique identifier for the document. */
+    fun token(): Optional<String> = Optional.ofNullable(token.getNullable("token"))
 
     /** Globally unique identifier for the account holder. */
     fun accountHolderToken(): Optional<String> =
@@ -43,15 +45,12 @@ private constructor(
     fun documentType(): Optional<DocumentType> =
         Optional.ofNullable(documentType.getNullable("document_type"))
 
-    /** Globally unique identifier for the entity. */
-    fun entityToken(): Optional<String> =
-        Optional.ofNullable(entityToken.getNullable("entity_token"))
-
+    /** List of required document images to upload. */
     fun requiredDocumentUploads(): Optional<List<RequiredDocumentUpload>> =
         Optional.ofNullable(requiredDocumentUploads.getNullable("required_document_uploads"))
 
     /** Globally unique identifier for the document. */
-    fun token(): Optional<String> = Optional.ofNullable(token.getNullable("token"))
+    @JsonProperty("token") @ExcludeMissing fun _token() = token
 
     /** Globally unique identifier for the account holder. */
     @JsonProperty("account_holder_token")
@@ -61,27 +60,21 @@ private constructor(
     /** Type of documentation to be submitted for verification. */
     @JsonProperty("document_type") @ExcludeMissing fun _documentType() = documentType
 
-    /** Globally unique identifier for the entity. */
-    @JsonProperty("entity_token") @ExcludeMissing fun _entityToken() = entityToken
-
+    /** List of required document images to upload. */
     @JsonProperty("required_document_uploads")
     @ExcludeMissing
     fun _requiredDocumentUploads() = requiredDocumentUploads
-
-    /** Globally unique identifier for the document. */
-    @JsonProperty("token") @ExcludeMissing fun _token() = token
 
     @JsonAnyGetter
     @ExcludeMissing
     fun _additionalProperties(): Map<String, JsonValue> = additionalProperties
 
-    fun validate(): AccountHolderDocument = apply {
+    fun validate(): AccountHolderSimulateEnrollmentDocumentReviewResponse = apply {
         if (!validated) {
+            token()
             accountHolderToken()
             documentType()
-            entityToken()
             requiredDocumentUploads().map { it.forEach { it.validate() } }
-            token()
             validated = true
         }
     }
@@ -93,12 +86,11 @@ private constructor(
             return true
         }
 
-        return other is AccountHolderDocument &&
+        return other is AccountHolderSimulateEnrollmentDocumentReviewResponse &&
+            this.token == other.token &&
             this.accountHolderToken == other.accountHolderToken &&
             this.documentType == other.documentType &&
-            this.entityToken == other.entityToken &&
             this.requiredDocumentUploads == other.requiredDocumentUploads &&
-            this.token == other.token &&
             this.additionalProperties == other.additionalProperties
     }
 
@@ -106,11 +98,10 @@ private constructor(
         if (hashCode == 0) {
             hashCode =
                 Objects.hash(
+                    token,
                     accountHolderToken,
                     documentType,
-                    entityToken,
                     requiredDocumentUploads,
-                    token,
                     additionalProperties,
                 )
         }
@@ -118,7 +109,7 @@ private constructor(
     }
 
     override fun toString() =
-        "AccountHolderDocument{accountHolderToken=$accountHolderToken, documentType=$documentType, entityToken=$entityToken, requiredDocumentUploads=$requiredDocumentUploads, token=$token, additionalProperties=$additionalProperties}"
+        "AccountHolderSimulateEnrollmentDocumentReviewResponse{token=$token, accountHolderToken=$accountHolderToken, documentType=$documentType, requiredDocumentUploads=$requiredDocumentUploads, additionalProperties=$additionalProperties}"
 
     companion object {
 
@@ -127,23 +118,36 @@ private constructor(
 
     class Builder {
 
+        private var token: JsonField<String> = JsonMissing.of()
         private var accountHolderToken: JsonField<String> = JsonMissing.of()
         private var documentType: JsonField<DocumentType> = JsonMissing.of()
-        private var entityToken: JsonField<String> = JsonMissing.of()
         private var requiredDocumentUploads: JsonField<List<RequiredDocumentUpload>> =
             JsonMissing.of()
-        private var token: JsonField<String> = JsonMissing.of()
         private var additionalProperties: MutableMap<String, JsonValue> = mutableMapOf()
 
         @JvmSynthetic
-        internal fun from(accountHolderDocument: AccountHolderDocument) = apply {
-            this.accountHolderToken = accountHolderDocument.accountHolderToken
-            this.documentType = accountHolderDocument.documentType
-            this.entityToken = accountHolderDocument.entityToken
-            this.requiredDocumentUploads = accountHolderDocument.requiredDocumentUploads
-            this.token = accountHolderDocument.token
-            additionalProperties(accountHolderDocument.additionalProperties)
+        internal fun from(
+            accountHolderSimulateEnrollmentDocumentReviewResponse:
+                AccountHolderSimulateEnrollmentDocumentReviewResponse
+        ) = apply {
+            this.token = accountHolderSimulateEnrollmentDocumentReviewResponse.token
+            this.accountHolderToken =
+                accountHolderSimulateEnrollmentDocumentReviewResponse.accountHolderToken
+            this.documentType = accountHolderSimulateEnrollmentDocumentReviewResponse.documentType
+            this.requiredDocumentUploads =
+                accountHolderSimulateEnrollmentDocumentReviewResponse.requiredDocumentUploads
+            additionalProperties(
+                accountHolderSimulateEnrollmentDocumentReviewResponse.additionalProperties
+            )
         }
+
+        /** Globally unique identifier for the document. */
+        fun token(token: String) = token(JsonField.of(token))
+
+        /** Globally unique identifier for the document. */
+        @JsonProperty("token")
+        @ExcludeMissing
+        fun token(token: JsonField<String>) = apply { this.token = token }
 
         /** Globally unique identifier for the account holder. */
         fun accountHolderToken(accountHolderToken: String) =
@@ -166,30 +170,16 @@ private constructor(
             this.documentType = documentType
         }
 
-        /** Globally unique identifier for the entity. */
-        fun entityToken(entityToken: String) = entityToken(JsonField.of(entityToken))
-
-        /** Globally unique identifier for the entity. */
-        @JsonProperty("entity_token")
-        @ExcludeMissing
-        fun entityToken(entityToken: JsonField<String>) = apply { this.entityToken = entityToken }
-
+        /** List of required document images to upload. */
         fun requiredDocumentUploads(requiredDocumentUploads: List<RequiredDocumentUpload>) =
             requiredDocumentUploads(JsonField.of(requiredDocumentUploads))
 
+        /** List of required document images to upload. */
         @JsonProperty("required_document_uploads")
         @ExcludeMissing
         fun requiredDocumentUploads(
             requiredDocumentUploads: JsonField<List<RequiredDocumentUpload>>
         ) = apply { this.requiredDocumentUploads = requiredDocumentUploads }
-
-        /** Globally unique identifier for the document. */
-        fun token(token: String) = token(JsonField.of(token))
-
-        /** Globally unique identifier for the document. */
-        @JsonProperty("token")
-        @ExcludeMissing
-        fun token(token: JsonField<String>) = apply { this.token = token }
 
         fun additionalProperties(additionalProperties: Map<String, JsonValue>) = apply {
             this.additionalProperties.clear()
@@ -205,13 +195,12 @@ private constructor(
             this.additionalProperties.putAll(additionalProperties)
         }
 
-        fun build(): AccountHolderDocument =
-            AccountHolderDocument(
+        fun build(): AccountHolderSimulateEnrollmentDocumentReviewResponse =
+            AccountHolderSimulateEnrollmentDocumentReviewResponse(
+                token,
                 accountHolderToken,
                 documentType,
-                entityToken,
                 requiredDocumentUploads.map { it.toUnmodifiable() },
-                token,
                 additionalProperties.toUnmodifiable(),
             )
     }
@@ -238,140 +227,53 @@ private constructor(
 
         companion object {
 
-            @JvmField val EIN_LETTER = DocumentType(JsonField.of("EIN_LETTER"))
+            @JvmField val COMMERCIAL_LICENSE = DocumentType(JsonField.of("commercial_license"))
 
-            @JvmField val TAX_RETURN = DocumentType(JsonField.of("TAX_RETURN"))
+            @JvmField val DRIVERS_LICENSE = DocumentType(JsonField.of("drivers_license"))
 
-            @JvmField val OPERATING_AGREEMENT = DocumentType(JsonField.of("OPERATING_AGREEMENT"))
+            @JvmField val PASSPORT = DocumentType(JsonField.of("passport"))
 
-            @JvmField
-            val CERTIFICATE_OF_FORMATION = DocumentType(JsonField.of("CERTIFICATE_OF_FORMATION"))
+            @JvmField val PASSPORT_CARD = DocumentType(JsonField.of("passport_card"))
 
-            @JvmField val DRIVERS_LICENSE = DocumentType(JsonField.of("DRIVERS_LICENSE"))
-
-            @JvmField val PASSPORT = DocumentType(JsonField.of("PASSPORT"))
-
-            @JvmField val PASSPORT_CARD = DocumentType(JsonField.of("PASSPORT_CARD"))
-
-            @JvmField
-            val CERTIFICATE_OF_GOOD_STANDING =
-                DocumentType(JsonField.of("CERTIFICATE_OF_GOOD_STANDING"))
-
-            @JvmField
-            val ARTICLES_OF_INCORPORATION = DocumentType(JsonField.of("ARTICLES_OF_INCORPORATION"))
-
-            @JvmField
-            val ARTICLES_OF_ORGANIZATION = DocumentType(JsonField.of("ARTICLES_OF_ORGANIZATION"))
-
-            @JvmField val BYLAWS = DocumentType(JsonField.of("BYLAWS"))
-
-            @JvmField
-            val GOVERNMENT_BUSINESS_LICENSE =
-                DocumentType(JsonField.of("GOVERNMENT_BUSINESS_LICENSE"))
-
-            @JvmField
-            val PARTNERSHIP_AGREEMENT = DocumentType(JsonField.of("PARTNERSHIP_AGREEMENT"))
-
-            @JvmField val SS4_FORM = DocumentType(JsonField.of("SS4_FORM"))
-
-            @JvmField val BANK_STATEMENT = DocumentType(JsonField.of("BANK_STATEMENT"))
-
-            @JvmField
-            val UTILITY_BILL_STATEMENT = DocumentType(JsonField.of("UTILITY_BILL_STATEMENT"))
-
-            @JvmField val SSN_CARD = DocumentType(JsonField.of("SSN_CARD"))
-
-            @JvmField val ITIN_LETTER = DocumentType(JsonField.of("ITIN_LETTER"))
+            @JvmField val VISA = DocumentType(JsonField.of("visa"))
 
             @JvmStatic fun of(value: String) = DocumentType(JsonField.of(value))
         }
 
         enum class Known {
-            EIN_LETTER,
-            TAX_RETURN,
-            OPERATING_AGREEMENT,
-            CERTIFICATE_OF_FORMATION,
+            COMMERCIAL_LICENSE,
             DRIVERS_LICENSE,
             PASSPORT,
             PASSPORT_CARD,
-            CERTIFICATE_OF_GOOD_STANDING,
-            ARTICLES_OF_INCORPORATION,
-            ARTICLES_OF_ORGANIZATION,
-            BYLAWS,
-            GOVERNMENT_BUSINESS_LICENSE,
-            PARTNERSHIP_AGREEMENT,
-            SS4_FORM,
-            BANK_STATEMENT,
-            UTILITY_BILL_STATEMENT,
-            SSN_CARD,
-            ITIN_LETTER,
+            VISA,
         }
 
         enum class Value {
-            EIN_LETTER,
-            TAX_RETURN,
-            OPERATING_AGREEMENT,
-            CERTIFICATE_OF_FORMATION,
+            COMMERCIAL_LICENSE,
             DRIVERS_LICENSE,
             PASSPORT,
             PASSPORT_CARD,
-            CERTIFICATE_OF_GOOD_STANDING,
-            ARTICLES_OF_INCORPORATION,
-            ARTICLES_OF_ORGANIZATION,
-            BYLAWS,
-            GOVERNMENT_BUSINESS_LICENSE,
-            PARTNERSHIP_AGREEMENT,
-            SS4_FORM,
-            BANK_STATEMENT,
-            UTILITY_BILL_STATEMENT,
-            SSN_CARD,
-            ITIN_LETTER,
+            VISA,
             _UNKNOWN,
         }
 
         fun value(): Value =
             when (this) {
-                EIN_LETTER -> Value.EIN_LETTER
-                TAX_RETURN -> Value.TAX_RETURN
-                OPERATING_AGREEMENT -> Value.OPERATING_AGREEMENT
-                CERTIFICATE_OF_FORMATION -> Value.CERTIFICATE_OF_FORMATION
+                COMMERCIAL_LICENSE -> Value.COMMERCIAL_LICENSE
                 DRIVERS_LICENSE -> Value.DRIVERS_LICENSE
                 PASSPORT -> Value.PASSPORT
                 PASSPORT_CARD -> Value.PASSPORT_CARD
-                CERTIFICATE_OF_GOOD_STANDING -> Value.CERTIFICATE_OF_GOOD_STANDING
-                ARTICLES_OF_INCORPORATION -> Value.ARTICLES_OF_INCORPORATION
-                ARTICLES_OF_ORGANIZATION -> Value.ARTICLES_OF_ORGANIZATION
-                BYLAWS -> Value.BYLAWS
-                GOVERNMENT_BUSINESS_LICENSE -> Value.GOVERNMENT_BUSINESS_LICENSE
-                PARTNERSHIP_AGREEMENT -> Value.PARTNERSHIP_AGREEMENT
-                SS4_FORM -> Value.SS4_FORM
-                BANK_STATEMENT -> Value.BANK_STATEMENT
-                UTILITY_BILL_STATEMENT -> Value.UTILITY_BILL_STATEMENT
-                SSN_CARD -> Value.SSN_CARD
-                ITIN_LETTER -> Value.ITIN_LETTER
+                VISA -> Value.VISA
                 else -> Value._UNKNOWN
             }
 
         fun known(): Known =
             when (this) {
-                EIN_LETTER -> Known.EIN_LETTER
-                TAX_RETURN -> Known.TAX_RETURN
-                OPERATING_AGREEMENT -> Known.OPERATING_AGREEMENT
-                CERTIFICATE_OF_FORMATION -> Known.CERTIFICATE_OF_FORMATION
+                COMMERCIAL_LICENSE -> Known.COMMERCIAL_LICENSE
                 DRIVERS_LICENSE -> Known.DRIVERS_LICENSE
                 PASSPORT -> Known.PASSPORT
                 PASSPORT_CARD -> Known.PASSPORT_CARD
-                CERTIFICATE_OF_GOOD_STANDING -> Known.CERTIFICATE_OF_GOOD_STANDING
-                ARTICLES_OF_INCORPORATION -> Known.ARTICLES_OF_INCORPORATION
-                ARTICLES_OF_ORGANIZATION -> Known.ARTICLES_OF_ORGANIZATION
-                BYLAWS -> Known.BYLAWS
-                GOVERNMENT_BUSINESS_LICENSE -> Known.GOVERNMENT_BUSINESS_LICENSE
-                PARTNERSHIP_AGREEMENT -> Known.PARTNERSHIP_AGREEMENT
-                SS4_FORM -> Known.SS4_FORM
-                BANK_STATEMENT -> Known.BANK_STATEMENT
-                UTILITY_BILL_STATEMENT -> Known.UTILITY_BILL_STATEMENT
-                SSN_CARD -> Known.SSN_CARD
-                ITIN_LETTER -> Known.ITIN_LETTER
+                VISA -> Known.VISA
                 else -> throw LithicInvalidDataException("Unknown DocumentType: $value")
             }
 
@@ -387,7 +289,6 @@ private constructor(
         private val status: JsonField<Status>,
         private val statusReasons: JsonField<List<StatusReason>>,
         private val uploadUrl: JsonField<String>,
-        private val token: JsonField<String>,
         private val additionalProperties: Map<String, JsonValue>,
     ) {
 
@@ -402,6 +303,7 @@ private constructor(
         /** Status of document image upload. */
         fun status(): Optional<Status> = Optional.ofNullable(status.getNullable("status"))
 
+        /** Reasons for document image upload status. */
         fun statusReasons(): Optional<List<StatusReason>> =
             Optional.ofNullable(statusReasons.getNullable("status_reasons"))
 
@@ -414,15 +316,13 @@ private constructor(
          */
         fun uploadUrl(): Optional<String> = Optional.ofNullable(uploadUrl.getNullable("upload_url"))
 
-        /** Globally unique identifier for the document upload. */
-        fun token(): Optional<String> = Optional.ofNullable(token.getNullable("token"))
-
         /** Type of image to upload. */
         @JsonProperty("image_type") @ExcludeMissing fun _imageType() = imageType
 
         /** Status of document image upload. */
         @JsonProperty("status") @ExcludeMissing fun _status() = status
 
+        /** Reasons for document image upload status. */
         @JsonProperty("status_reasons") @ExcludeMissing fun _statusReasons() = statusReasons
 
         /**
@@ -434,9 +334,6 @@ private constructor(
          */
         @JsonProperty("upload_url") @ExcludeMissing fun _uploadUrl() = uploadUrl
 
-        /** Globally unique identifier for the document upload. */
-        @JsonProperty("token") @ExcludeMissing fun _token() = token
-
         @JsonAnyGetter
         @ExcludeMissing
         fun _additionalProperties(): Map<String, JsonValue> = additionalProperties
@@ -447,7 +344,6 @@ private constructor(
                 status()
                 statusReasons()
                 uploadUrl()
-                token()
                 validated = true
             }
         }
@@ -464,7 +360,6 @@ private constructor(
                 this.status == other.status &&
                 this.statusReasons == other.statusReasons &&
                 this.uploadUrl == other.uploadUrl &&
-                this.token == other.token &&
                 this.additionalProperties == other.additionalProperties
         }
 
@@ -476,7 +371,6 @@ private constructor(
                         status,
                         statusReasons,
                         uploadUrl,
-                        token,
                         additionalProperties,
                     )
             }
@@ -484,7 +378,7 @@ private constructor(
         }
 
         override fun toString() =
-            "RequiredDocumentUpload{imageType=$imageType, status=$status, statusReasons=$statusReasons, uploadUrl=$uploadUrl, token=$token, additionalProperties=$additionalProperties}"
+            "RequiredDocumentUpload{imageType=$imageType, status=$status, statusReasons=$statusReasons, uploadUrl=$uploadUrl, additionalProperties=$additionalProperties}"
 
         companion object {
 
@@ -497,7 +391,6 @@ private constructor(
             private var status: JsonField<Status> = JsonMissing.of()
             private var statusReasons: JsonField<List<StatusReason>> = JsonMissing.of()
             private var uploadUrl: JsonField<String> = JsonMissing.of()
-            private var token: JsonField<String> = JsonMissing.of()
             private var additionalProperties: MutableMap<String, JsonValue> = mutableMapOf()
 
             @JvmSynthetic
@@ -506,7 +399,6 @@ private constructor(
                 this.status = requiredDocumentUpload.status
                 this.statusReasons = requiredDocumentUpload.statusReasons
                 this.uploadUrl = requiredDocumentUpload.uploadUrl
-                this.token = requiredDocumentUpload.token
                 additionalProperties(requiredDocumentUpload.additionalProperties)
             }
 
@@ -526,9 +418,11 @@ private constructor(
             @ExcludeMissing
             fun status(status: JsonField<Status>) = apply { this.status = status }
 
+            /** Reasons for document image upload status. */
             fun statusReasons(statusReasons: List<StatusReason>) =
                 statusReasons(JsonField.of(statusReasons))
 
+            /** Reasons for document image upload status. */
             @JsonProperty("status_reasons")
             @ExcludeMissing
             fun statusReasons(statusReasons: JsonField<List<StatusReason>>) = apply {
@@ -555,14 +449,6 @@ private constructor(
             @ExcludeMissing
             fun uploadUrl(uploadUrl: JsonField<String>) = apply { this.uploadUrl = uploadUrl }
 
-            /** Globally unique identifier for the document upload. */
-            fun token(token: String) = token(JsonField.of(token))
-
-            /** Globally unique identifier for the document upload. */
-            @JsonProperty("token")
-            @ExcludeMissing
-            fun token(token: JsonField<String>) = apply { this.token = token }
-
             fun additionalProperties(additionalProperties: Map<String, JsonValue>) = apply {
                 this.additionalProperties.clear()
                 this.additionalProperties.putAll(additionalProperties)
@@ -583,7 +469,6 @@ private constructor(
                     status,
                     statusReasons.map { it.toUnmodifiable() },
                     uploadUrl,
-                    token,
                     additionalProperties.toUnmodifiable(),
                 )
         }
@@ -671,7 +556,7 @@ private constructor(
 
                 @JvmField val FAILED = Status(JsonField.of("FAILED"))
 
-                @JvmField val PENDING_UPLOAD = Status(JsonField.of("PENDING_UPLOAD"))
+                @JvmField val PENDING = Status(JsonField.of("PENDING"))
 
                 @JvmField val UPLOADED = Status(JsonField.of("UPLOADED"))
 
@@ -681,14 +566,14 @@ private constructor(
             enum class Known {
                 COMPLETED,
                 FAILED,
-                PENDING_UPLOAD,
+                PENDING,
                 UPLOADED,
             }
 
             enum class Value {
                 COMPLETED,
                 FAILED,
-                PENDING_UPLOAD,
+                PENDING,
                 UPLOADED,
                 _UNKNOWN,
             }
@@ -697,7 +582,7 @@ private constructor(
                 when (this) {
                     COMPLETED -> Value.COMPLETED
                     FAILED -> Value.FAILED
-                    PENDING_UPLOAD -> Value.PENDING_UPLOAD
+                    PENDING -> Value.PENDING
                     UPLOADED -> Value.UPLOADED
                     else -> Value._UNKNOWN
                 }
@@ -706,7 +591,7 @@ private constructor(
                 when (this) {
                     COMPLETED -> Known.COMPLETED
                     FAILED -> Known.FAILED
-                    PENDING_UPLOAD -> Known.PENDING_UPLOAD
+                    PENDING -> Known.PENDING
                     UPLOADED -> Known.UPLOADED
                     else -> throw LithicInvalidDataException("Unknown Status: $value")
                 }