+++ lithic-java-core/src/test/kotlin/com/lithic/api/models/AccountHolderUploadDocumentParamsTest.kt
@@ -12,7 +12,8 @@ class AccountHolderUploadDocumentParamsTest {
     fun createAccountHolderUploadDocumentParams() {
         AccountHolderUploadDocumentParams.builder()
             .accountHolderToken("182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e")
-            .documentType(AccountHolderUploadDocumentParams.DocumentType.COMMERCIAL_LICENSE)
+            .documentType(AccountHolderUploadDocumentParams.DocumentType.EIN_LETTER)
+            .entityToken("182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e")
             .build()
     }
 
@@ -21,12 +22,14 @@ class AccountHolderUploadDocumentParamsTest {
         val params =
             AccountHolderUploadDocumentParams.builder()
                 .accountHolderToken("182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e")
-                .documentType(AccountHolderUploadDocumentParams.DocumentType.COMMERCIAL_LICENSE)
+                .documentType(AccountHolderUploadDocumentParams.DocumentType.EIN_LETTER)
+                .entityToken("182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e")
                 .build()
         val body = params.getBody()
         assertThat(body).isNotNull
         assertThat(body.documentType())
-            .isEqualTo(AccountHolderUploadDocumentParams.DocumentType.COMMERCIAL_LICENSE)
+            .isEqualTo(AccountHolderUploadDocumentParams.DocumentType.EIN_LETTER)
+        assertThat(body.entityToken()).isEqualTo("182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e")
     }
 
     @Test
@@ -34,12 +37,9 @@ class AccountHolderUploadDocumentParamsTest {
         val params =
             AccountHolderUploadDocumentParams.builder()
                 .accountHolderToken("182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e")
-                .documentType(AccountHolderUploadDocumentParams.DocumentType.COMMERCIAL_LICENSE)
                 .build()
         val body = params.getBody()
         assertThat(body).isNotNull
-        assertThat(body.documentType())
-            .isEqualTo(AccountHolderUploadDocumentParams.DocumentType.COMMERCIAL_LICENSE)
     }
 
     @Test
@@ -47,7 +47,6 @@ class AccountHolderUploadDocumentParamsTest {
         val params =
             AccountHolderUploadDocumentParams.builder()
                 .accountHolderToken("182bd5e5-6e1a-4fe4-a799-aa6d9a6ab26e")
-                .documentType(AccountHolderUploadDocumentParams.DocumentType.COMMERCIAL_LICENSE)
                 .build()
         assertThat(params).isNotNull
         // path param "accountHolderToken"