+++ lithic-java-core/src/test/kotlin/com/lithic/api/models/StatementsTest.kt
@@ -20,7 +20,7 @@ class StatementsTest {
                             .accountStanding(
                                 Statement.AccountStanding.builder()
                                     .periodNumber(123L)
-                                    .state(Statement.AccountStanding.AccountState2.STANDARD)
+                                    .periodState(Statement.AccountStanding.PeriodState.STANDARD)
                                     .build()
                             )
                             .amountDue(123L)
@@ -72,7 +72,7 @@ class StatementsTest {
                     .accountStanding(
                         Statement.AccountStanding.builder()
                             .periodNumber(123L)
-                            .state(Statement.AccountStanding.AccountState2.STANDARD)
+                            .periodState(Statement.AccountStanding.PeriodState.STANDARD)
                             .build()
                     )
                     .amountDue(123L)